Bug #45310 repl: Simplify rename handling
With Bug #34355 fixed with git:07d8c9a42eb7c25234387c611a164337c60f3dce
since release-3.2-2~116 UDL guarantees to always perform the two parts
of a rename/move transaction in sequence. Therefor it is sufficient to
keep the state in memory and the code to persist it to disk can (and
should) be removed completely.