Bug #51368 UCR: Fix ucr.clear()
collecction.abc.MutableMapping.clear() is implemented to do a
repetitive pop() until the mapping is clear:

/usr/lib/python3.7/_collections_abc.py:816
> def clear(self):
>     try:
>         while True:
>             self.popitem()
>     except KeyError:
>         pass

This does not work with UCR and its layers, as only the writeable layer
is drained, but other layers may still have values: A new view is
created after each iteration, which also includes the keys from the
other layers. If one of those is pop()ed, KeyError is risen and the
loop aborts too early.

Also this is very inefficient as for each iteration a new merged view
has to be created.

Implement clear().