Make univention-base-* Priority: standard
`debootstrp` by default installs all `Priority: required+important`
packages, which also pulls in `univention-base-packages`. Through a long
dependency chain
* cron →
* default-mta[V] →
* univention-mail-postfix →
* python-univention-directory-manager →
* python-univention-lib →
* python-samba →
* libtracker-sparql-2.0-0 →
* libsoup2.4-1 →
* glib-networking →
* gsettings-desktop-schemas →
* dconf-gsettings-backend →
* dconf-service →
* dbus-session-bus[V]
is pulled in, which are mostly `Priority: optional` and are not prepared
to be pulled in that early. `dbus-sesstion-bus` is a pure virtual
package, provided by `dbus-x11` or `dbus-user-session`. `debootstrap` is
not prepared to resolve that dependency via a virtual package and
installation fails.

Don't make `univention-mail-postfix` `Provide: default-mta`.

Change to Priority to not install univention-base-* during debootstrap
time.