Bug #18501: Fix handling of proxyAddresses mapping
Fix handling of proxyAddresses mapping in combination with
the Diff-Mode code:

We need to compare mapped values in sync_from_ucs, because:

1. We must use mapped values in case a mapping function is defined,
   not only for single_value attributes.

2. proxyAddresses is constructed from mailPrimaryAddress *and*
mailAlternativeAddress. The first is marked by capital "SMTP:"
prefix in AD, while the other is marked by lowercase "smtp:".

3. If only mailAlternativeAddress is changed, then still to_proxyAddresses
needs to be called to construct the new proxyAddresses value, but
that mapping function is only attached to mailPrimaryAddress.

Before Diff-Mode, the sync_from_ucs method iterated over attributes
of the mapped object. I think that makes sense, so we can deal only
with mapped values in sync_from_ucs. So I adjusted the method
to get the mapped object_old as argument, instead of the 'old' dict
from the listener. Then it also doesn't need the 'new' dict either,
making the code more readable.