Bug #51567: Revert last commit and make a simpler fix
Allow snyc_from_ucs synchronization model for otherHomePhone
to coexist with that for proxyAddresses in sync_from_ucs.

If we iterate over UCS ldap_attribute values then changes to
mailAlternativeAddress alone don't trigger the to_proxyAddresses
function which is connected to the primaryMailAddress attribute.

So, to support this special replication mechanism for proxyAddresses
we have to iterate over the mapped (AD) attributes.

Caveat: If connector/ad/mapping/user/primarymail=true (default)
the mapping for mailAlternativeAddress is also considered, but
in that case it is configured to `read` mode, so there is no
additional modification in sync_from_ucs.