Bug #42287 pkgdb: Break package dependency cycle
Make the UMC package only "Recommend: univention-pkgdb" as the later is
also the meta package, which is installed when the component is
installed.

Fix the UMC module to also work on servers differently then the database
server by using the machine account as a fallback if the UNIX socket
fails.

Remove the legacy packet.

Fix debian/rules to not execute shell code multiple times.

Remove old cruft from debian/rules.

Improve package descriptions in debian/control.