Bug #51607: allow use of the self service without reentering credentials
The self service could only be used by entering the credentials as form values.
Now, when being logged in into UMC via SAML or via plain login, these
credentials can be used.

In environments where the self-service is installed on a DC Master and
on e.g. a DC Slave one need to consider, that all UCR variables,
settings e.g. adjustments of the email texts now need to be set on the DC
Slave as well. As this is a breaking change the behavior must be enabled
explicitly by setting 'umc/self-service/allow-saml' to true.

Conceptionally this change would have required the UMC UDM module to be
installed on the DC Slave. That's not possible. Therefore the method
get_user_attributes had to be split into get_user_attributes_values
which is executed locally while get_user_attributes_descriptions must be
executed on the DC Master.

Sorry for this worst commit I ever did in the last time.