Bug #49978 l10n: Fix binary_package_name naming
The variable is named wrong, because it is initialized with the *source*
package name from the current working directory path, which does now
work when the package is built from a temporary directory.

Re-name it to "package" and use the value from "module_name", which
already actually is the "binary package name".

Merge the now common code into
UMCModuleTranslation._read_module_attributes_from_source_package()