Use univention-join sequence for DebHelper
TODO: not all replacements did move the #DEBHELPER# block to the
previous call_joinscript position. check carefully before applying hunks
from this patch!

if UCR variables are set in the postinst which are used in the
joinscript we should probably replace the ucr set calls with UCR
default variable definitions.