Bug #51588 up: Fix apt-mirror clean
The apt-mirror `clean` statement expects only a single argument: an URL
which is internally converted to a path by stripping the schema, host
and port prefix and by normalizing the path. This is then used as a
starting point to crawl that directory locally and to compare is with
the list of expected files `ALL`.

Before UCS-5 we used the "flat layout", e.g.
> deb https://updates.software-univention.de/4.0/maintained/ 4.0-0/all/
so for each $major.$minor release and per [un]maintained we had multiple
starting points.

With UCS-5 we use the "pool" layout, where both "dists/" and "pool/" are
directly below the base:
> deb https://updates.software-univention.de/ ucs500 main
Therefore there its sufficient to give "clean" a single URL as the
starting point.

Fixes: 0f37d183f0f057e14f091c29c25987a81bde862e