Bug #52194 lib/i18n: Fix setting locale
The regular expression does not implement whats written in the comment:
"territory", "codeset and "modifier" are all optional according to
<man:setlocale(3)>:
> language[_territory][.codeset][@modifier]

The "C" and "POSIX" locales may also have a "codeset" and "modifier".

Add PEP 484 Python static type annotation.

Use the empty string instead of None for type safety.

Reverts: c74670fd4a8f3465fc1da21237fa8711763d9a90

PS: The is no `:charmap` in the locale spec: this is an artifact from
UCS/UCR as <man:locale.gen(5)> always requires a `charmap`, but
`.codeset` is optional. Therefore it must always be specified explicitly
in `/etc/locale.gen`.
UCS stores this in UCRV `locale` and `locale/default`, which are space
separated lists. Therefore `:` was chosen to separate the parts of the
2-tuple (locale, charmap).