Bug #52043: Fix bind start with samba4 backend
The named daemon could not open the control channel:
`open: /etc/bind/rndc.key: permission denied`
even though the daemon runs as root.
It is necessary to start bind as root to work with samba. Changing the
permission on `rndc.key` is not a stable option either:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=386791

The named daemon now starts as root but with the group bind, which seems
to work fine with both samba and the `rndc.key` file.