Bug #43686: replace intid.d with systemctl for univention-directory-listener
coode like
if [ -x /etc/init.d/univention-directory-listener ]; then
    invoke-rc.d univention-directory-listener crestart
is replaced by
if [ -n "$(systemctl list-unit-files --type=service --state enabled,diasbled\
        --no-legend univention-directory-listener.service)" ]; then
    systemctl try-restart univention-directory-listener

so that the restart of the service is only tried if the service exists
and if it is already running.

code like

test -x /etc/init.d/univention-directory-listener && invoke-rc.d univention-directory-listener restart

is simplified by

systemctl restart univention-directory-listener

because it makes no difference here if the service not exists, or
could not be restarted.