Bug #52274: run UMC-Server with addditional UNIX socket
Start the UMC-Server with an additional UNIX socket and let the
UMC-Webserver connect to this.
Therefor we prevent the long runnning TLS handshake between the two
services which we trust anyway.
We must keep the TCP port 6670 currently opened because there are a few
umc-command calls which still connect remotely to it.