Bug #52316: fix faulty connections and prevent lots of log messages
384 def _connection(self, socket):
385     '''Signal callback: Invoked on incoming connections.'''
386     try:
387         socket, addr = socket.accept()
388     except EnvironmentError as exc:
389         CORE.error('Cannot accept new connection: %s' % (exc,))
390         soft, hard = resource.getrlimit(resource.RLIMIT_NOFILE)
391         resource.setrlimit(resource.RLIMIT_NOFILE, (soft + 2, hard + 2))
392         try:
393             socket, addr = socket.accept()
394             socket.close()
395         except EnvironmentError:
396             pass
397         finally:
398             resource.setrlimit(resource.RLIMIT_NOFILE, (soft, hard))
399         return True
400     fd = socket.fileno()
401     flags = fcntl.fcntl(fd, fcntl.F_GETFD)
402     fcntl.fcntl(fd, fcntl.F_SETFD, flags | fcntl.FD_CLOEXEC)

The code shown here should actually deal with the case that not enough free
file descriptors are available (→ errno 24: Too many open files).
Since the listening socket has a backlog (for 100 connections →
SERVER_MAX_CONNECTIONS), incoming TCP connections are accepted and established
by the kernel, but not yet passed from the kernel to the UMC server process.

If there are no more free file descriptors available, the socket.accept() call
always throws an errno 24 error and the already established TCP connections
starve to death in the backlog, unless the UMC server gets free file descriptors
again. To prevent starvation, the code temporarily raises the limit for open
filehandles by 2, executes the accept() and immediately closes the connection
again. This causes the UMC web server not to get a valid connection to the UMC
server and returns an error 503 to the browser, so that the browser is informed
and the request does not wait infinitely for an accept().


In the new multi-process UMC server, all UMC servers have their listening socket
bound to the same port. As long as there is nothing to do, all UMC server
processes wait in epoll().
When a new connection comes in, all processes are woken up and they call the
_connection() method, where a socket.accept() is called.
In the following we assume that UMCServer A has successfully accepted the new
socket. All other processes (B,C,D,...) get an EnvironmentError exception with
errno 11, because there is nothing to do at the moment.
(Errno 11 stands for "Resource temporarily unavailable"/"The resource is
currently not available" or EAGAIN).

The above code does not distinguish between errno 11 and errno 24 and from time
to time a special corner case occurs e.g. with process B:
between line 387 (first accept()) and line 393 (second accept()) a new TCP
connection from the UMC web server comes in. The code for "no free filehandles"
accepts the new TCP connection in line 393 and closes it again immediately in
line 394. As described above, the UMC web server then returns an error message
503, but this time there was no error at all and the connection was immediately
closed again in an unjustified way.

The fix for this problem now catches errno 11 before the log output in line 389
and immediately makes a "return True" to avoid unnecessary log output.
Furthermore, lines 390 to 398 are now only executed on errno 24.

For all other errors, a log message and a traceback is logged, to be able
to debug these occurences.