Bug #52314: do not disable session polling during portal usage
We need to detect when the SAML session is timed out so that we can make
a new passive saml login in the iframe so users 1. don't get logged out
and 2. even if they are logged out, the logout button is stilled
rendered and when one clicks on it the user is already logged out from
UMC so only a page refresh is done which then does the passive SAML
iframe login and one is logged in again.