Bug #52321: reduce polling of get/session-info
When 10000 users are logged in sumultanously we have in average 333 session-info requests per second.
Currently we are doing every 30 seconds a get/session-info request which asks for the session validity
and checks if one is still logged in. If not one is redirected to the passive iframe SAML login to
update the SAML assertion, which is by default only valid for 5 minutes, and is required for LDAP and PAM authentication at the UMC-Server.

So we reduce the polling to the time when the session times out, and in
case the server didn't yet destroy the session in the next 30 seconds
the polling is done every 30 seconds.