Bug #52338: do not store relay state between requests
It is not necessary to store a dictionary relay_state and custom data in
the users Session between the SAML AuthNRequest and the response to it
to redirect the user back to the location where he came from.

Instead we should just set the URL/path as relay state (this is often seen SAML behavior).

Then we don't leak a never removed dictionary entry in case the user
never comes back (and we currently don't removed it at all ever).