Bug #51461: Fix welcome screen
Move the installation of univention-grub and univention-welcome-screen
into the debian installer through a dependency from the
univention-system-setup.

Otherwise the libdrm2 dependency for plymouth is missing because it was
earlier removed in the setup by an `apt autoremove` call. This also
ensures that the kernel is using our parameters already in the
univention-system-setup.