Bug #51508 fix error in setup_saml_sp.py:
Problem:
setup_saml_sp.py fails during the package setup which results in not all
templates being committed.

Python2:
urlparse did throw an `AttributeError` if `umc/saml/idp-server` was not set.
The `AttributeError` was beeing caught (Bug #52434) and the module
execution was successful with a warning.

Python3:
urlparse now works with `None` and does not throw an `AttributeError`
anymore. Instead ElementTree now throws a `FileNotFoundError` which is
not caught and the module execution fails.

Fix:
Exit early if `umc/saml/idp-server` is not set.