Bug #52763: create consistent config and restart named processes
It looks like the named config is not consistent at the time the
joinscript is called. The files univention.conf and/or
univention.conf.proxy are missing and therefore named is unable to
start.
Additionally, if the named has not been running before, up to this
change the named is not (re)started by the join script, and this
could lead to one/two non-running named processes.