Bug #52797: fix translations
xgettext doesn't seem to extract text from shell scripts if the first
argument to gettext isn't the phrase to translate. I'm not sure how this
had worked in the past, but it finally broke with the changes to the
univention-l10n package.

I also added translations for the status text under the univention logo.
Plymount script doesn't natively support translations, so I have to
inject the text through a shell script.