Bug #52798: fix syntax error during UCS 4.4 -> 5.0 upgrade
Our patch to grub in UCS 4.4 for /etc/grub.d/05_debian_theme creates a
syntax error if no color is set in /etc/default/grub. As a workaround
one color is set, it has no further effect after the update is complete
and the workaround can be removed with 5.1

Setting a package relationship did not work because neither "Breaks" nor
"Conflicts" controls the configuration order of packages. A "Depends" is
problematic during the system setup, because the debian installer
installs and configures grub as the final package. It should not be
installed any sooner.

error: syntax error.
error: Incorrect command.
error: syntax error.
Syntax error at line 114
Syntax errors are detected in generated GRUB config file.