fix[ucr]: Re-implement default layer
the main problem was that the class must show different behavior
depending on the user:
- for `ucs update` we need to access the *raw* values
- for `ucr get` we need to query the *resolved* values

Instead of using a sub-class add the code for resolving in the 2
locations for `get()` and `_merge()`.

Remove sub-class _DefaultConfigRegistry

For `_merge()` this also solves a performance problem as
`_DefaultConfigRegistry.items()` would resolve *all* values *every time*,
even when the value is unneeded as when set in a higher layer.

Bug #38938