fix[ucr]: Call handlers with old&new visible value
We must also get the *old* value of the previously highest layer. This
is complicated by the fact, that `_run_changed()` is called *after* the
change was already performed, so we manually search the layers *below*
for the first one having a value. (Searching the layers *above* cam be
skipped as then the higher layer is still in place and the following
test with `if reg > ucr.scope` will trigger.)

So we have these cases:

1. Adding a new overwrite needs to get the old value from the layer
below and the new value from the current change, which will be the
newest highest layer.

2. Change the value on the highest layer will use the old and new value
from this layer.

3. Removing an overwrite will use the old value from the change and the
new value from the then highest new layer.

Bug #52847
Fixes: e589e6b7b519038cfd5d2cae9e19f36032511fbb
Fixes: 6535f1408969b8b071244c769f48ef05ffbac177