fix[ucr]: Call handlers only for changed variables
BREAKING CHANGE:
When UCR variables are changed the list of changed variables (from the
command line or to `update_ucr()`) were passed unfiltered to the
internal function  `_run_changed()`, which does the grunt work of
calling the registered handlers.
This even happened for UCR variables shadowed by layers of higher
priority, leading to handlers being called for variables, which did not
change visibly. Even worse if the used the `new` valued passed via
`changed` they would use the wrong "shadowed" value:

   ucr set --forced key=forced  # changed={"key", (None, "forced")}
   ucr set          key=normal  # changed={"key", (None, "normal")}

Only a message was printed, that the change was shadowed.
A handler would need to use `ucr.get(key)` to get the real value instead
of suing `new` directly.
This strange behavior is at best unexpected and all UCR scripts /
handlers in UCR either don't care about the values at all or expect the
new value.

The new behavior filters the list of changes and only passes on those,
which are *not* shadowed, e.g where the change happens in the layer with
the current highest priority (in the example from above the 2nd `set`
would *not* trigger any handler.)

Doing a null-change (setting a UCR variable to its current value again)
will still trigger the handlers as this work-around is often used to
re-commit files like `/etc/apt/sources.list.d/`.

Bug #52847