Bug #52881: use different docker image for install tests
start-test.sh now checks for `docker-image:` in the cfg file
and uses this image in docker mode.

This is useful for vnc install tests. We have an extra docker
image for the vnc tools (based on ucs-ec2-tools + some
vnc libs/tools) and with this setup we can define the image
in the cfg file itself (the defaut ist still the image in
start-test.sh).