Bug #47734: python3 compatibility for selenium Test 17_self_service_registration
The smtpd.SMTPServer got some major changes while switching to python3. for once giving it an empty address causes problems as soon as the system also provides ipv6 support on that port. therefore changing it to localhost for the tests. The method process_message of this lib got the kwargs while switching to python3. Since messages now arrive in bytes format the mail content hast to be decoded