Bug #52923: disable fetchmail during the update
In the postinst script of fetchmail, the fetchmail service is restarted
automatically during the update to UCS 5.0-0. If there is no "poll" entry
in /etc/fetchmailrc, the restart of the fetchmail polling daemon fails
and unfortunately the fetchmail.postinst also fails. This causes an
abortion of the update to UCS 5.0-0.

To prevent this problematic behaviour, in preup.sh the fetchmail.service
stopped and then disabled via the corresponding UCR variable:

  ucr set fetchmail/autostart=no

The old value of "fetchmail/autostart" is saved in the UCR variable
"fetchmail/autostart/update500" and restored in postup.sh. Also a restart
of fetchmail.service is attempted in postup.sh.