Bug #52836: Fixed `Failed to open /dev/tty` in ucs-test
This has already been commited for the 4.4 branch, see
    https://git.knut.univention.de/univention/ucs/-/commit/db19441ee3dc20096a631dcc6039ca7241a060c2

Original text:

The file /usr/lib/python2.7/dist-packages/univention/testing/data.py
contains the code to run() our tests. It uses `Popen` with
`preexec_fn=os.setsid`, which is a questionable solution, because
setsid also detaches the process from the controlling tty*1 when
detaching the process from its parent. This causes the isatty() function
to return a false positive and that is what systemd uses to detect
if the channels have to receive an EOF*2. That call fails with a
well-known error message `Failed to open /dev/tty`.
We cannot expect the code to be removed from systemd, because from my
understanding it is there for the same reason why the python manual
claims about subprocess.call, that it would be disadvised to use
stderr=PIPE with it.

*1 see the section `Getting rid of a controlling tty` here:
https://www.win.tue.nl/~aeb/linux/lk/lk-10.html

*2 https://github.com/systemd/systemd/blob/main/src/basic/process-util.c

Bug #52836: Fixed `Failed to open /dev/tty [...]` in tests

Python 3 migration done