fix[net]: Move UCR setting to DHCP script
etc/network/if-up.d/ scripts are only called ONCE during the initial
network start.
DHCP setting on the other hand can change every time a lease is renewed.

Move the scripts from /etc/network/if-up.d/ to
/etc/dhcp/dhclient-exit-hooks.d/.
Convert and combine them into a single shell script as required by the
DHClient API.

Merge the code for storing the current address in UCR with the code for
UMC.
Please note: this removes a forced restart of UMC.

DNS is handled by /etc/dhcp/dhclient-enter-hooks.d/resolvconf already,
so remove the code for that.

Bug #52959