fix[net]: Depend on apache2.service
The IP address is registered using the UMC HTTP API at the Primary node.

On the Primary itself this requires apache2, umc, and umc-web to be
already functional, otherwise the unit will fail. As
`univention-register-network-address` uses HTTPS this requires
apache2.service, which itself pulls in umc-web and that umc, so
apache2.service is enough.

On all other roles that dependency is not necessary and only delays the
registration of the IP address, which is mostly optional anyway.

Bug #52959