fix[uss]: Prevent USS from starting
on initial installation and on upgrades: It only ever should be started
on reboot.

The services must be `enabled` to be pulled in by `multi-user.target`.
But the service must not be (re-)started on installation or on upgrade.

`dh_installsystemd` has the annoying behavior that it does 2 things:
- it installs the files debian/$pkg[.$name].$suffix, but it can not
  handle multiple files for the same package.
- this requires using the option `--name`.
- but in its second phase is generates {pre,post}{rm,inst} snippets for
  all *already* installed units, which leads to duplicate snippets.

Manually install the unit files via `dh_install`.
Only use `dh_installsystemd` to generate the snippets.

Bug #52977
Fixes: fedae5c5ba3b37b2018f344305736470f8fa3b41