fix[grub]: Close gap between Debian SHIM and Univention GRUB2
Since we switched to SHIM, GRUB2 and Linux kernel from Debian there is a
problem with `grub-install`: It uses `GRUB_DISTRIBUTOR=Univention,..`
from `/etc/default/grub` as the `distribution_id`, which is used for
installing `grub.cfg` and `*.efi` to `/boot/efi/EFI/univention/`, but
Debians GRUB2 has `.../debian/` burned into the signed binary.

Therefore we have to place a dummy their to fill to boot chain gap.

Thanks go to Sebastian Mohr <sebi@univention.mibes.it> for providing the
information.

PS: As an alternative you can do
> grub-install --target=x86_64-efi --bootloader-id debian -v
to force the installer to ignore `GRUB_DISTRIBUTOR`, but their is no
easy switch to do that always. It falls back to config.grub_distributor,
which is used in many more locations. Probably we still want
"Univention" to show up instead of "Debian", so changing it does not
seem to be a good option.

Bug #51984