Bug #50648: fix creation of pointer records for IPv6 computers
ADMIN       ( ERROR   ) : Could not determine host record for name='wc87f7zlia', ip='2001:1:2:3::2'. Not creating pointer record.

When searching for the aAAAARecord (host record) a search filter with the short form 2001:1:2:3::2 was made instead of the long/expanded form.
The best solution would be with a IP address matching rule, but OpenLDAP has none.
As UDM is always saving host records in the expanded form, this is okay for now and we don't need a software comparision.