Bug #53175: Restart DHCP server after join
Installation of univention-dhcp fails on a non-primary systems already
joined, which will be the norm with UCS 5.0-0 as USS no longer installs
DHCP as part of the initial installation.

The problem is that
  univention-dhcp: Depends isc-dhcp-server-ldap: Depends isc-dhcp-server

/var/lib/dpkg/info/isc-dhcp-server.postinst tries to start the daemon
with /etc/dhcp/dhcpd.conf already templated by univention-dhcp.
This will fail as the daemon will do a LDAP search first for
  (&(objectClass=dhcpServer)(cn=$hostname))
which is only created later by
/usr/lib/univention-install/25univention-dhcp.inst when later it gets
invoked from /var/lib/dpkg/info/univention-dhcp.postinst.

The failed start will abort isc-dhcp-server.postinst, which aborts the
full package installation.

We have augmented isc-dhcp-server with
/usr/lib/systemd/system/isc-dhcp-server.service.d/univention-dhcp.conf:
  ConditionPathExists=/etc/dhcp/dhcpd.conf
  ConditionPathExists=/var/univention-join/joined
which are both fulfilled as the system is already joined and the file exists.

Add a second flag file which is only created
- on upgrades
- after the join script has been xecuted