fix[udm]: Remove one of multiple DNS PTR RRs
> Traceback (most recent call last):
...
>   File "/usr/lib/python3/dist-packages/univention/admin/handlers/computers/ipmanagedclient.py", line 219, in _ldap_post_modify
>     univention.admin.handlers.simpleComputer._ldap_post_modify(self)
>   File "/usr/lib/python3/dist-packages/univention/admin/handlers/__init__.py", line 2717, in _ldap_post_modify
>     self.__remove_dns_reverse_object(self['name'], None, entry)
>   File "/usr/lib/python3/dist-packages/univention/admin/handlers/__init__.py", line 2285, in __remove_dns_reverse_object
>     rdn = self.calc_dns_reverse_entry_name(ip, zone)
>   File "/usr/lib/python3/dist-packages/univention/admin/handlers/__init__.py", line 2905, in calc_dns_reverse_entry_name
>     assert rev.endswith(subnet)
> AssertionError

In __remove_dns_reverse_object() performans an LDAP search by name (e.g.
(pTRRecord=$hostname)) and returns *all* DNS PTR RRs.  This will return
both IPv4 and IPv6 RRs - even multiple if the host has multiple IPv4 or
IPv6 addresses.  For each zone calc_dns_reverse_entry_name() is called,
even when the IP address is not in that zone.  In that case the assert()
was raised.

Convert it to raising a ValueError().

Bug #53213