feat(udm-rest): support application/json-patch+json documents
PATCH method should be used with patch documents.
The CLI client needs a way to set values by adding, replacing or
removing them. But the client doesn't know the data strucutre.
Therefore these values are given as patch document to the server which
can then compare the values for equality (e.g. after normalization of
DNs)