fix[kvm]: Remove all old network configuration
The interface name even within our KVM environment is not stable: An one
case the interface was named "ens7" instead of "ens3" and was left
untouched by the cleanup script leading to the IP surviving the
templating step: Afterwards the old IP address was reused as the DNS
server.

Just remove all UCRV related to the network configuration.

Bug #51967