ci: Use debian/changelog time-stamp
`git checkout` uses $NOW. Two builds from git thus have different
time-stamps.
For normal Debian builds this is not a problem as there the files are
only checked out once and the time-stamps are preserved by the
intermediate TAR file.

Use the time of the latest debian/changelog entry for all files.

Bug #53108