Bug #52174: Check if more cpu cores increase stability
In one of the last tests the system-stats.log showed a "load average" > 1,
while a test failed with a network error.
I think it is worth a try to check if more cores can decrease our number
of flaky tests due to timeouts.

I will revert this commit after starting one test run.