Bug #53181: fix lost stdout from ssh
In case univention-join, which uses univention-ssh, was called from the
join umc module, the stdout was lost from the ssh call. The remote ucs
version could not be retrieved and the join wasn't possible.

I'm not sure what exactly caused this. It seems to be some buffer flush
problem with ssh being forked in a non interactive shell.