test[utils.sh]: Force DHCP
dnsmasq4kvm does not always provide a IPv4 address in time, in which
case the VMs falls back to the IPv4-LL address. Even when the correct
address is obtained later on the IPv4-LL remains in UCR - at least with
UCS-4.4 - with UCS-5.0 this is supposed to be fixed by Bug #44462.

Reset timeout back to default 30s to give dhclient more time and tries
to obtain the IPv4 address, which is set by test/utils/base_appliance.sh

Disable the IPv4-LL fallback mechanism.

Bug #52174