fix[test/base]: Disable 104_check-archive-keyring
The test was broken until release-5.0-0-902-g77b06aac33:
All our 4.x-x KVM images are broken and contain the Debian package
`debian-archive-keyring`, which they should not.

The first fixed image is `5.0-0`.

But the upgrade tests will start with `4.4-8` and will still have the
package installed, so disabling it for just `5.0-0` does  not work.

Disable the test for now until `5.0-1` is released.

Bug #53768
Fixes: 0b0748d99c4