Bug #54339: allow to search for dns/ptr_record ip=$address
The property "ip" of a dns/ptr_record is not part of the mapping and therefore not evaluated in a filter.
Therefor the filter needs to be manually rewritten in the UDM module.

This is not quite easy as the IP address is not part of the dns/ptr_records LDAP attributes - only the address (relativeDomainName) and the subnet reverse address (zoneName). The IP is constructed virtually.

So we have to search for the host parts of the given IP address only in the reverse zones matching the subnet.
But we don't know the subnet - the IP address in your filter doesn't contain a subnet mask.

And in real life there could be more subnet masks as the common /8 and /16 and /24 and 32.
We could search for every possible combination - but this would be a large LDAP filter.