WIP Bug #52964: fix error handling during move operation
When errors happen during moving objects there is only a flag "errors=True"
given to the client when fetching progress. This is never evaluated the
progress is fetched forever. We should respond with an HTTP error
instead to indicate an error.

Example:
"UDM_Error: This operation is not allowed on this object. Objects of the
"dhcp/service" object type can not be moved."

TODO: test if it works in multiprocessing mode