Bug #54418: Convert bibliorefs to translatable footnotes
First tries with the Sphinx bibliography feature didn't work. Therefore,
I decided to use RST footnotes. The text looks similar to the
bibliography style in Docbook and most important, the link text and the
link itself shows up in the translation. And the links are collected at
the buttom of the document.