Bug #54504: Remove wrong code for replication.py priority
The u-d-l package in UCS 4.4 has an error and does not hardcode the
priority for replication.py correctly. The strcmp call checks the loaded
modules file basename, while the function gets the absolute path of the
file. As a fix, set the priority to 0.0 directly in the listener module,
as it is done in UCS 5.