Bug #54472: Remove references to UCS 4.3
The text had three references to UCS 4.3 errata updates that showed when
the features were introduced to the App Center. Since UCS 4.3 is out of
maintenance, the references were removed. We assume that users operate
UCS with a maintained version so that the required errata level is given
for the functionality.