Bug #54418: Docs pipeline: Use en only when no other locales exist
Documentation that is not localized and only available in English, does
not have a locales directory with any other languages. In this case, the
build-ci script would use "*" for one language. This resulted in a
broken generated-config-doc.yml.

When a documentation does not have a locales directory, the iteration
continues with next step and ignores locales/*.