Bug #54523: Add HTML with generic redirect
The HTML file replaces the DocBook single page documentation. It
redirects the visitor automatically to the new documentation location
after 15 seconds and uses the HTTP meta refresh. It also shows a
message about the redirection and a link to click, if the redirection
does not work automatically.

For testing purposes, it has localhost configured as host for the
moment. This needs to be changed later.