Bug #54523: Add anchor mapping for redirects
The docbook single page uses anchor tags to refer to section headings
and figures. To redirect them, JavaScript is needed. This change adds a
mapping from the anchor tags to the new locations of the documentation.

For testing it currently has localhost configured as base URL.

Furthermore, missing reference labels have been added to RST files.