Bug #54523: Pipeline exclude files in HTML build from sync
The Sphinx HTML build process generates files that cause problems in the
final link check and that are not needed in the deployment. One of those
files is _static/webpack-macros.html which is excluded from being copied
to the public directory.

Another file not needed in the deployment is .buildinfo.