Bug #54536: Upgrade Univention Sphinx Extension v0.0.11
Sphinx offers the setting html_baseurl which is responsible for the
canonical URL in the HTML page header and relevant to search engines and
sitemaps.

Without the adoption to v0.0.11 the canonical URLs will disappear for
App Center and Quickstart. The setting explicitly activates the logic
for the correct URL setting provided by the extension.