Bug #54536: Adjust links to UCS manual
The docs `app-center`, `architecture` and `quickstart` use default
external links to the UCS manual. Since the manual is transformed to
RST, those links break the Sphinx linkcheck and thus the pipeline. They
work for the reader.

This change adjusts the links and uses the intersphinx feature with
label references and let Sphinx resolve the links automatically.