Bug #44997: remove locking options
The per share options to configure oplocks (& locks) don't seem useful to me.
If a particular customer has oplock issues, you would probably disable them fully as a temporary workaround and not on a per share basis.
Also this is not a common option to configure, which needs to be accessible conveniently via UMC.

Bug #7843 says:
> Die Locking-Optionen sollten meines Erachtens überhaupt nicht explizit konfigurierbar sein:
>
> Wenn der "Samba-Schreibzugriff" aus dem "Samba-Rechte"-Reiter deaktiviert ist, gibt es keine konkurrierenden Schreibzugriffe. Dann kann das Locking automatisch deaktiviert und Fake-Oplocks aktiviert werden.
>
> Wenn konkurrierende Schreibzugriffe möglich sind, gibt es keinen vernünftigen Grund Locking zu deaktivieren (und wer das aus irgendwelchen Gründen doch möglich, kann das immer noch über "Erweiterte Samba-Einstellungen" machen), von daher sollten die Locking-Optionen besser komplett raus.