Bug #7843: remove "Block size"
This parameter controls the behavior of smbd(8) when reporting disk free sizes. By default, this reports a disk block size of 1024 bytes.

Changing this parameter may have some effect on the efficiency of client writes, this is not yet confirmed. This parameter was added to allow advanced
administrators to change it (usually to a higher value) and test the effect it has on client write performance without re-compiling the code. As this is an
experimental option it may be removed in a future release.

Changing this option does not change the disk free reporting size, just the block size unit reported to the client.