Bug #54737: Pipeline: Cleanup after docs production job
The production job collects the docs build artifacts and commits them
for publication. Unfortunately, the job doesn't cleanup afterwards. When
files don't exist in the build artifacts anymore, they stay in the
publication repository because they are still in the production job file
collection.

To fix this, the production job cleans up the directory where it
collects all the files, after the scripts doing the job are finished.