Bug #52881: Fix spelling error in German Quickstart Guide
Commit 83e8a4ced9a6ae24cf31f0cb0d5dce18458548df changed the index.rst of
the quickstart guide, but the translation file was not updated. The
affected section is not translated anymore and appears in English
language in the spell check for German. The spell check fails in the
pipeline.

To fix it, I ran:

```
$ cd doc/quickstart
$ make gettext
$ sphinx-intl update -p _build/gettext/ -l de
```

I checked the `locales/de/LC_MESSAGES/index.po` file for `fuzzy` entries
and fixed them. I committed the updated file.