Bug #TODO: schoolserver join often fails in big environments
The s4connector join script waits for the synchronization of the dns objects,
as well as some kerberos principals. If a replica server is joined, all replicated objects
were added as pickle files before. This means that the necessary objects for the join are only
added to samba4 last, which can take several hours in big environments.
Resyncing them first, makes the join faster, and avoids customers having to restart the join
if they ran into the timeout.