Bug #54748: make sure Python 2.7 UDM modules are written again by the listener
As the Python 2.7 modules were installed directly in
/usr/lib/python2.7/dist-packages/univention/admin/handlers/ms/ but also
registered by the listener and we now removed the files from the
packages we have to trigger the listener again. By just increasing the
joinscript version the registration will update the package version and
trigger the listener so that those files are re-written into the correct
location.

Fixes: 97d76baf2a55c5fc976a626c4ce2d0a251bedf18