Bug #54797: udn: Build index at start
Building the index at the start of the execution will improve
notably the performance of notify_transcation_get_one_dn when
the transaction file is big.
Change notify_transcation_get_one_dn to start reading from the
end of the file when the transaction is not found in the index.