Bug #54883: move operational attributes into a private dictionary
The UCS@school Exam mode copied all attributes from `oldattr` and
created a new exam user with those attributes. This fails because
`entryUUID` is part of it, while the LDAP server only allows to set this
via relax control.

For now, we put the operational attributes into a private instance.
UCS@school has to exclude them explicitly. As soon as this is ensured
we can move the operational attributes back into oldattr, so they can be
accessed from outside as well.