ci: Switch to merge-request pipeline
So far we have used Branch-pipelines, which run for every branch
- release branch
- feature branch
- tag

GitLab's `rules:changes:` mechanism uses the previous commit `HEAD~` to
detect which paths were touched and which jobs needs to be scheduled.
For a new branch `HEAD~` is empty, which triggers a complete rebuild of
*all* jobs.
This is undesirable for branch tests, where only *changed* packages
should be re-built and tested.

Merge-Request-pipelines on the other hand are only triggered if a merge
request is created or updated: it is associated with a source and target
GIT branch, which are used instead to calculate the changed paths: this
includes all accumulated changes between those two branches, not only
the files touched by the last commit.
This is very handy to re-build only the packages changed by the branch
associated to merge request.
One minor issue is that all touched packages are re-built on every new
commit added to the branch, even if the last commit only touches one
package. As most feature branches only touch one package this is only an
issue for those few merge requests, which touch many packages. If this
becomes a problem re-building the packages can be cached.

Switch to GitLab Container registry.

Add Kaniko caching to improve building `ucslint` on every push.

Issue univention/dist/repo-ng#107